+++ CHANGELOG.md
@@ -101,6 +101,7 @@ And (*as always*) many bugfixes and improvements to existing features!
 - **Other**
   - **prompt enhance** add `allura-org/Gemma-3-Glitter-4B` model support  
   - **prompt enhance** add `Qwen/Qwen3-4B-Instruct-2507` model support  
+  - **caption** add `Qwen/Qwen2.5-VL-3B-Instruct` model support  
   - **schedulers** add **Flash FlowMatch**  
 - **API**
   - add `/sdapi/v1/checkpoint` POST endpoint to simply load a model